Remove rawlsModel dependency from workbenchMetrics
