+++ metrics/src/main/scala/org/broadinstitute/dsde/rawls/metrics/Expansion.scala
@@ -1,10 +1,7 @@
 package org.broadinstitute.dsde.rawls.metrics
 
 import java.util.UUID
-
-import org.broadinstitute.dsde.rawls.model.{RawlsEnumeration, WorkspaceName}
 import spray.http.{HttpMethod, StatusCode, Uri}
-
 import scala.annotation.implicitNotFound
 
 /**
@@ -25,25 +22,9 @@ object Expansion {
   // Typeclass instances:
 
   /**
-    * Implicit expansion for WorkspaceName.
-    * Statsd doesn't allow slashes in metric names, so we override makeName to override
-    * the default toString based implementation.
-    */
-  implicit object WorkspaceNameExpansion extends Expansion[WorkspaceName] {
-    override def makeName(n: WorkspaceName): String = n.toString.replace('/', '.')
-  }
-
-  /**
     * Implicit expansion for UUID using the default makeName.
     */
   implicit object UUIDExpansion extends Expansion[UUID]
-
-  /**
-    * Implicit expansion for RawlsEnumeration using the default makeName.
-    * This takes an upper type bound {{{A <: RawlsEnumeration}}} so it can work with any
-    * subtype of RawlsEnumeration.
-    */
-  implicit def RawlsEnumerationExpansion[A <: RawlsEnumeration[_]] = new Expansion[A] {}
 
   /**
     * Implicit expansion for HttpMethod.
